feat: changed button color and updated product context
